Getting initial files and class setup.
Got parts of the parser that should be in sniffer pulled apart. It compiles with this as a static class. I got a template constructor and members for the dialect.

Make it so it is file line ending independent! no longer cares if it is '\r\n', '\r', or '\n'. some test files names changed because of this.

just some small todos for future me.

add some changes to put in the has header functionality. It doesn't work yet but the structure is there.